upgrading ios, android and javascript to RN  0.31.0
