+++ package.json
@@ -3,25 +3,33 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "start": "node node_modules/react-native/local-cli/cli.js start"
+    "start": "node node_modules/react-native/local-cli/cli.js start",
+    "clean:babelrc": "find ./node_modules -name react-packager -prune -o -name '.babelrc' -print | xargs rm -f",
+    "postinstall": "npm run clean:babelrc"
   },
   "dependencies": {
     "immutable": "3.8.1",
-    "react": "15.0.2",
-    "react-native": "0.26.0",
+    "react": "15.3.0",
+    "react-native": "0.31.0",
     "react-native-keyboard-spacer": "0.2.0",
     "react-redux": "4.4.5",
     "redux": "3.5.2",
     "redux-localstorage": "0.4.0",
     "redux-logger": "2.6.1",
     "redux-thunk": "2.0.1",
-    "reselect": "2.5.1"
+    "reselect": "2.5.1",
+    "react-proxy": "1.1.2",
+    "react-deep-force-update": "2.0.1"
   },
   "devDependencies": {
     "babel-eslint": "6.0.4",
     "eslint": "2.9.0",
     "eslint-loader": "1.3.0",
     "eslint-plugin-react": "5.0.1",
-    "rnpm-plugin-upgrade": "0.26.0"
+    "mocha": "^2.4.5",
+    "react-addons-test-utils": "15.3.0",
+    "react-dom": "15.3.0",
+    "react-native-mock": "0.2.5",
+    "sinon": "^1.17.4"
   }
 }
